nitpick comments
